+++ tests/ProdManager/helpers/test_helpers_hello.py
@@ -1,4 +1,4 @@
-from pkg_resources import parse_version
+from packaging.version import parse as parse_version
 import pytest
 
 import responses